# ReservationAppClient/docs/api-docs.yaml
openapi: 3.1.0
info:
  title: OpenAPI definition
  version: v0
servers:
  - url: http://localhost:8080
    description: Generated server url
paths:
...
  /api/venue/{id}:
    get:
      tags:
        - venue-controller
      operationId: getVenue
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/VenueDto"
    put:
      tags:
        - venue-controller
      operationId: editVenue
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                venue:
                  $ref: "#/components/schemas/VenueDto"
                file:
                  type: string
                  format: binary
              required:
                - venue
      responses:
        "204":
          description: No Content
    delete:
      tags:
        - venue-controller
      operationId: deleteVenue
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "204":
          description: No Content
...
  /api/court/{id}:
    get:
      tags:
        - court-controller
      operationId: showCourt
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CourtDto"
    put:
      tags:
        - court-controller
      operationId: editCourt
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                court:
                  $ref: "#/components/schemas/CourtDto"
                file:
                  type: string
                  format: binary
              required:
                - court
      responses:
        "204":
          description: No Content
    delete:
      tags:
        - court-controller
      operationId: deleteCourt
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "204":
          description: No Content
...
  /api/venue:
    get:
      tags:
        - venue-controller
      operationId: getAllVenues
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/VenueDto"
    post:
      tags:
        - venue-controller
      operationId: createVenue
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                venue:
                  $ref: "#/components/schemas/VenueDto"
                file:
                  type: string
                  format: binary
              required:
                - venue
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/VenueDto"
...
  /api/court:
    get:
      tags:
        - court-controller
      operationId: showAllCourts
      parameters:
        - name: courtFilter
          in: query
          required: true
          schema:
            $ref: "#/components/schemas/CourtFilter"
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CourtDto"
    post:
      tags:
        - court-controller
      operationId: createCourt
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                court:
                  $ref: "#/components/schemas/CourtDto"
                file:
                  type: string
                  format: binary
              required:
                - court
      responses:
        "201":
          description: Created
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CourtDto"
...
components:
  schemas:
...
    CourtDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
          minLength: 1
        surface:
          type: string
          enum:
            - CLAY
            - HARD
            - WIMBLEDON
        indoor:
          type: boolean
        lighting:
          type: boolean
        venue:
          $ref: "#/components/schemas/VenueDto"
        photoUrl:
          type: string
      required:
        - indoor
        - lighting
        - name
        - surface
        - venue
...
    VenueDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
          minLength: 1
        address:
          type: string
          minLength: 1
        phoneNumber:
          type: string
          maxLength: 2147483647
          minLength: 9
        photoUrl:
          type: string
      required:
        - address
        - name
        - phoneNumber
...
    CourtFilter:
      type: object
      properties:
        venueId:
          type: integer
          format: int64
        name:
          type: string
        surface:
          type: string
          enum:
            - CLAY
            - HARD
            - WIMBLEDON
        indoor:
          type: boolean
        lighting:
          type: boolean
...